requirements.txt
pillow>=9.0.0
requests>=2.30.0
opencv-python
onnxruntime-gpu
numpy
tqdm